# docs/api/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OmniFlow FastAPI 接口设计",
    "version": "1.0.0-draft",
    "description": "设计契约第一版，与Spec v1.0配套。尚未部署、不是从现有server.py或运行中的FastAPI生成；不得用文档成功代替真实接口验收。"
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "待实现的相对前缀；不表示现有域名已提供这些接口"
    }
  ],
...
  "security": [
    {
      "SessionCookie": []
    }
  ],
  "paths": {
...
    "/tasks": {
      "post": {
        "operationId": "create_task",
        "summary": "通过统一服务创建一个媒体任务",
        "tags": [
          "任务"
        ],
        "description": "每task一份媒体。精修入口显式提交，Gemini经内部受控网关调用同一任务服务；参数和确认事实必须由后端校验。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "202": {
            "description": "通过统一服务创建一个媒体任务",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              },
              "Idempotency-Replayed": {
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "409": {
            "$ref": "#/components/responses/HTTP409"
          },
          "410": {
            "$ref": "#/components/responses/HTTP410"
          },
          "422": {
            "$ref": "#/components/responses/HTTP422"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          },
          "507": {
            "$ref": "#/components/responses/HTTP507"
          }
        },
        "x-implementation-status": "design-only",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "list_tasks",
        "summary": "查询自己的媒体任务",
        "tags": [
          "任务"
        ],
        "description": "查询自己的媒体任务",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageCursor"
          },
          {
            "$ref": "#/components/parameters/PageLimit"
          },
          {
            "$ref": "#/components/parameters/ConversationFilter"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "queued",
                "submitting",
                "running",
                "saving",
                "completed",
                "failed",
                "canceled",
                "submission_unknown",
                "needs_reconciliation"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询自己的媒体任务",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only"
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "operationId": "get_task",
        "summary": "查询真实媒体状态",
        "tags": [
          "任务"
        ],
        "description": "查询真实媒体状态",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询真实媒体状态",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "404": {
            "$ref": "#/components/responses/HTTP404"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only"
      }
    },
    "/tasks/{task_id}/cancel": {
      "post": {
        "operationId": "cancel_task",
        "summary": "仅取消尚未提交的媒体任务",
        "tags": [
          "任务"
        ],
        "description": "只允许queued→canceled，与worker领取原子互斥；其他非可取消状态409 TASK_NOT_CANCELABLE，绝不伪称供应商已撤销。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "仅取消尚未提交的媒体任务",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "404": {
            "$ref": "#/components/responses/HTTP404"
          },
          "409": {
            "$ref": "#/components/responses/HTTP409"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only"
      }
    },
    "/tasks/{task_id}/recover": {
      "post": {
        "operationId": "recover_task",
        "summary": "仅续查或重新下载同一供应商结果",
        "tags": [
          "任务"
        ],
        "description": "不新建task、不再次生成。有可核对供应商任务ID/结果线索才可恢复；提交未知且无线索409 RECONCILIATION_REQUIRED。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "仅续查或重新下载同一供应商结果",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "404": {
            "$ref": "#/components/responses/HTTP404"
          },
          "409": {
            "$ref": "#/components/responses/HTTP409"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only"
      }
    },
...
    "/admin/tasks": {
      "get": {
        "operationId": "admin_list_operational_tasks",
        "summary": "查看不含私密内容的任务故障概况",
        "tags": [
          "管理"
        ],
        "description": "查看不含私密内容的任务故障概况 必须校验admin角色；Cookie鉴权本身不代表管理员授权。",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageCursor"
          },
          {
            "$ref": "#/components/parameters/PageLimit"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "queued",
                "submitting",
                "running",
                "saving",
                "completed",
                "failed",
                "canceled",
                "submission_unknown",
                "needs_reconciliation"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查看不含私密内容的任务故障概况",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "服务器生成的追踪ID"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "受保护响应使用private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalTaskPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/HTTP400"
          },
          "401": {
            "$ref": "#/components/responses/HTTP401"
          },
          "403": {
            "$ref": "#/components/responses/HTTP403"
          },
          "429": {
            "$ref": "#/components/responses/HTTP429"
          },
          "500": {
            "$ref": "#/components/responses/HTTP500"
          },
          "503": {
            "$ref": "#/components/responses/HTTP503"
          }
        },
        "x-implementation-status": "design-only",
        "x-required-role": "admin"
      }
    },
...
  },
  "components": {
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]*$"
          },
          "detail": {
            "type": "string"
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "retryable": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "field",
                "code",
                "message"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "request_id",
          "retryable"
        ],
        "additionalProperties": false,
        "examples": [
          {
            "type": "about:blank",
            "title": "资源不可见",
            "status": 404,
            "code": "RESOURCE_NOT_FOUND",
            "detail": "资源不存在或当前用户不可访问。",
            "request_id": "22222222-2222-4222-8222-222222222222",
            "retryable": false
          }
        ]
      },
...
      "ResourceError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "additionalProperties": false
      },
...
      "ImageTaskCreate": {
        "type": "object",
        "properties": {
          "conversation_id": {
            "type": "string",
            "format": "uuid"
          },
          "target_artifact_id": {
            "type": "string",
            "format": "uuid"
          },
          "base_version_id": {
            "type": "string",
            "format": "uuid"
          },
          "regenerate_from_task_id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "const": "image"
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4000
          },
          "aspect_ratio": {
            "type": "string",
            "enum": [
              "1:1",
              "3:4",
              "4:3",
              "9:16",
              "16:9",
              "2:3",
              "3:2",
              "21:9"
            ]
          },
          "size_tier": {
            "type": "string",
            "enum": [
              "1K",
              "2K",
              "3K",
              "4K"
            ]
          },
          "reference_version_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "maxItems": 5,
            "uniqueItems": true
          }
        },
        "required": [
          "conversation_id",
          "kind",
          "prompt",
          "aspect_ratio",
          "size_tier"
        ],
        "additionalProperties": false,
        "dependentRequired": {
          "target_artifact_id": [
            "base_version_id"
          ],
          "base_version_id": [
            "target_artifact_id"
          ]
        },
        "examples": [
          {
            "conversation_id": "11111111-1111-4111-8111-111111111111",
            "kind": "image",
            "prompt": "冷白背景的虚构透明香水瓶，银色瓶盖与淡蓝丝带，无品牌文字。",
            "aspect_ratio": "9:16",
            "size_tier": "1K",
            "reference_version_ids": []
          }
        ]
      },
...
      "TaskCreate": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/ImageTaskCreate"
          },
          {
            "$ref": "#/components/schemas/VideoTaskCreate"
          },
          {
            "$ref": "#/components/schemas/LocalMotionTaskCreate"
          }
        ],
        "discriminator": {
          "propertyName": "kind",
          "mapping": {
            "image": "#/components/schemas/ImageTaskCreate",
            "ai_video": "#/components/schemas/VideoTaskCreate",
            "local_motion": "#/components/schemas/LocalMotionTaskCreate"
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "conversation_id": {
            "type": "string",
            "format": "uuid"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "type": "string",
            "enum": [
              "image",
              "ai_video",
              "local_motion"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "submitting",
              "running",
              "saving",
              "completed",
              "failed",
              "canceled",
              "submission_unknown",
              "needs_reconciliation"
            ]
          },
          "requested_parameters": {
            "$ref": "#/components/schemas/TaskCreate"
          },
          "execution_engine": {
            "type": "string",
            "enum": [
              "agnes-image-2.5-flash",
              "agnes-video-2.5-flash",
              "local-ffmpeg"
            ]
          },
          "output_version_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "can_cancel": {
            "type": "boolean"
          },
          "can_recover": {
            "type": "boolean"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceError"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "conversation_id",
          "run_id",
          "kind",
          "status",
          "requested_parameters",
          "execution_engine",
          "output_version_ids",
          "can_cancel",
          "can_recover",
          "error",
          "created_at",
          "updated_at"
        ],
        "additionalProperties": false,
        "examples": [
          {
            "id": "55555555-5555-4555-8555-555555555555",
            "conversation_id": "11111111-1111-4111-8111-111111111111",
            "run_id": null,
            "kind": "ai_video",
            "status": "queued",
            "requested_parameters": {
              "conversation_id": "11111111-1111-4111-8111-111111111111",
              "kind": "ai_video",
              "mode": "keyframe",
              "reference_confirmation_id": "88888888-8888-4888-8888-888888888888",
              "prompt": "保持参考图瓶体和文字不变，固定镜头，仅让丝带末端轻轻摆动。",
              "seconds": 4,
              "size_tier": "720P",
              "aspect_ratio": "9:16"
            },
            "execution_engine": "agnes-video-2.5-flash",
            "output_version_ids": [],
            "can_cancel": true,
            "can_recover": false,
            "error": null,
            "created_at": "2026-09-08T08:00:00Z",
            "updated_at": "2026-09-08T08:00:00Z"
          }
        ]
      },
...
      "TaskPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "additionalProperties": false
      },
...
      "OperationalTaskPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OperationalTask"
            }
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "additionalProperties": false
      },
...
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "同一用户动作重试必须复用；同键不同语义请求409。不是身份凭据。",
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9._:-]+$"
        }
      },
      "PageCursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 512
        }
      },
      "PageLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "ConversationFilter": {
        "name": "conversation_id",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
      "HTTP400": {
        "description": "HTTP 400；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "HTTP429": {
        "description": "HTTP 429；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          },
          "Retry-After": {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "description": "安全限流或队列背压；不是用户日配额"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "HTTP500": {
        "description": "HTTP 500；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "HTTP503": {
        "description": "HTTP 503；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "HTTP403": {
        "description": "HTTP 403；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "HTTP422": {
        "description": "HTTP 422；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "HTTP401": {
        "description": "HTTP 401；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "HTTP409": {
        "description": "HTTP 409；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "HTTP410": {
        "description": "HTTP 410；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "HTTP404": {
        "description": "HTTP 404；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "HTTP507": {
        "description": "HTTP 507；稳定错误码与处理方式见接口文档",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "服务器生成的追踪ID"
          },
          "Cache-Control": {
            "schema": {
              "type": "string"
            },
            "description": "受保护响应使用private, no-store"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-omniflow_session",
        "description": "网站登录，不是Google/Agnes凭据"
      },
      "CsrfContextCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-omniflow_csrf",
        "description": "匿名表单上下文，不具有登录权限"
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "绑定Cookie上下文，所有状态修改请求必须验证"
      },
      "MediaGrantToken": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "仅供应商取参考图片；后端生成短期高熵凭据，绑定grant_id、版本、任务和到期时间；不是网站或Google登录"
      }
    }
  },
...
}